Fixed incorrect command registration
