+++ tradingcards-plugin/src/main/java/net/tinetwork/tradingcards/tradingcardsplugin/TradingCards.java
@@ -11,16 +11,16 @@ import net.tinetwork.tradingcards.api.model.Rarity;
 import net.tinetwork.tradingcards.api.model.Series;
 import net.tinetwork.tradingcards.api.model.schedule.Mode;
 import net.tinetwork.tradingcards.tradingcardsplugin.card.TradingCard;
-import net.tinetwork.tradingcards.tradingcardsplugin.commands.BuySubCommand;
+import net.tinetwork.tradingcards.tradingcardsplugin.commands.BuyCommand;
 import net.tinetwork.tradingcards.tradingcardsplugin.commands.CardsCommand;
 import net.tinetwork.tradingcards.tradingcardsplugin.commands.CreateCommand;
 import net.tinetwork.tradingcards.tradingcardsplugin.commands.DebugCommands;
 import net.tinetwork.tradingcards.tradingcardsplugin.commands.DeckCommand;
 import net.tinetwork.tradingcards.tradingcardsplugin.commands.EditCommand;
 import net.tinetwork.tradingcards.tradingcardsplugin.commands.GiveCommands;
-import net.tinetwork.tradingcards.tradingcardsplugin.commands.ListSubCommand;
-import net.tinetwork.tradingcards.tradingcardsplugin.commands.MigrateSubCommand;
-import net.tinetwork.tradingcards.tradingcardsplugin.commands.SellSubCommand;
+import net.tinetwork.tradingcards.tradingcardsplugin.commands.ListCommand;
+import net.tinetwork.tradingcards.tradingcardsplugin.commands.MigrateCommand;
+import net.tinetwork.tradingcards.tradingcardsplugin.commands.SellCommand;
 import net.tinetwork.tradingcards.tradingcardsplugin.commands.edit.EditCard;
 import net.tinetwork.tradingcards.tradingcardsplugin.commands.edit.EditPack;
 import net.tinetwork.tradingcards.tradingcardsplugin.commands.edit.EditRarity;
@@ -286,12 +286,12 @@ public class TradingCards extends TradingCardsPlugin<TradingCard> {
         commandManager.registerCommand(new CardsCommand(this, playerBlacklist));
         commandManager.registerCommand(new EditCommand(this));
         commandManager.registerCommand(new CreateCommand(this));
-        commandManager.registerCommand(new BuySubCommand(this));
+        commandManager.registerCommand(new BuyCommand(this));
         commandManager.registerCommand(new DebugCommands(this));
         commandManager.registerCommand(new GiveCommands(this));
-        commandManager.registerCommand(new ListSubCommand(this));
-        commandManager.registerCommand(new MigrateSubCommand(this));
-        commandManager.registerCommand(new SellSubCommand(this));
+        commandManager.registerCommand(new ListCommand(this));
+        commandManager.registerCommand(new MigrateCommand(this));
+        commandManager.registerCommand(new SellCommand(this));
         commandManager.registerCommand(new DeckCommand(this));
         commandManager.enableUnstableAPI("help");
         commandManager.enableUnstableAPI("brigadier");
