+++ tradingcards-plugin/src/main/java/net/tinetwork/tradingcards/tradingcardsplugin/commands/GiveCommands.java
@@ -1,6 +1,7 @@
 package net.tinetwork.tradingcards.tradingcardsplugin.commands;
 
 import co.aikar.commands.BaseCommand;
+import co.aikar.commands.annotation.CommandAlias;
 import co.aikar.commands.annotation.CommandCompletion;
 import co.aikar.commands.annotation.CommandPermission;
 import co.aikar.commands.annotation.Default;
@@ -21,8 +22,7 @@ import org.bukkit.entity.Player;
 /**
  * @author sarhatabaot
  */
-@Subcommand("give")
-@CommandPermission(Permissions.GIVE)
+@CommandAlias("cards")
 public class GiveCommands extends BaseCommand {
     private final TradingCards plugin;
 
@@ -30,112 +30,117 @@ public class GiveCommands extends BaseCommand {
         this.plugin = plugin;
     }
 
-    @Subcommand("card")
-    @CommandPermission(Permissions.GIVE_CARD)
-    @Description("Gives a card.")
-    public class CardSubCommand extends BaseCommand {
-        @Default
-        @Description("Gives yourself a card.")
-        public void onDefault(final Player player, @Single final String rarity, @Single final String cardName) {
-            onPlayer(player, player.getName(), rarity, cardName, false);
+    @Subcommand("give")
+    @CommandPermission(Permissions.GIVE)
+    public class GiveSubCommands extends BaseCommand {
+
+        @Subcommand("card")
+        @CommandPermission(Permissions.GIVE_CARD)
+        @Description("Gives a card.")
+        public class CardSubCommand extends BaseCommand {
+            @Default
+            @Description("Gives yourself a card.")
+            public void onDefault(final Player player, @Single final String rarity, @Single final String cardName) {
+                onPlayer(player, player.getName(), rarity, cardName, false);
+            }
+
+            @Subcommand("player")
+            @CommandPermission(Permissions.GIVE_CARD_PLAYER)
+            @CommandCompletion("@players @rarities @cards")
+            public void onPlayer(final CommandSender sender, @Single final String playerName, @Single final String rarity, @Single final String cardName, @Single final boolean shiny) {
+                TradingCard card = plugin.getCardManager().getCard(cardName, rarity, shiny);
+                if (shiny && !card.hasShiny()) {
+                    ChatUtil.sendPrefixedMessage(sender, "This card does not have a shiny version.");
+                    return;
+                }
+
+                if (card instanceof EmptyCard) {
+                    ChatUtil.sendPrefixedMessage(sender, plugin.getMessagesConfig().noCard());
+                    return;
+                }
+
+                Player target = Bukkit.getPlayerExact(playerName);
+                if (isOnline(target)) {
+                    ChatUtil.sendPrefixedMessage(sender, CardsCommand.PLAYER_NOT_ONLINE);
+                    return;
+                }
+
+
+                ChatUtil.sendPrefixedMessage(target, plugin.getMessagesConfig().giveCard()
+                        .replace("%player%", target.getName())
+                        .replace("%card%", rarity + " " + cardName));
+
+                target.getInventory().addItem(card.build(shiny));
+            }
+
+            @Subcommand("shiny")
+            @CommandPermission(Permissions.GIVE_CARD_SHINY)
+            @CommandCompletion("@rarities @cards")
+            @Description("Gives a shiny card.")
+            public void onShiny(final Player player, @Single final String rarity, @Single final String cardName) {
+                onPlayer(player, player.getName(), rarity, cardName, true);
+            }
         }
 
-        @Subcommand("player")
-        @CommandPermission(Permissions.GIVE_CARD_PLAYER)
-        @CommandCompletion("@players @rarities @cards")
-        public void onPlayer(final CommandSender sender, @Single final String playerName, @Single final String rarity, @Single final String cardName, @Single final boolean shiny) {
-            TradingCard card = plugin.getCardManager().getCard(cardName, rarity, shiny);
-            if (shiny && !card.hasShiny()) {
-                ChatUtil.sendPrefixedMessage(sender, "This card does not have a shiny version.");
-                return;
-            }
-
-            if (card instanceof EmptyCard) {
-                ChatUtil.sendPrefixedMessage(sender, plugin.getMessagesConfig().noCard());
-                return;
-            }
-
-            Player target = Bukkit.getPlayerExact(playerName);
-            if (isOnline(target)) {
+        @Subcommand("pack")
+        @Description("Gives a pack to a player.")
+        @CommandCompletion("@players @packs")
+        @CommandPermission(Permissions.GIVE_PACK)
+        public void onGiveBoosterPack(final CommandSender sender, @Single final String playerName, @Single final String pack) {
+            Player player = Bukkit.getPlayerExact(playerName);
+            if (isOnline(player)) {
                 ChatUtil.sendPrefixedMessage(sender, CardsCommand.PLAYER_NOT_ONLINE);
                 return;
             }
 
+            CardUtil.dropItem(player, plugin.getPackManager().getPackItem(pack));
 
-            ChatUtil.sendPrefixedMessage(target, plugin.getMessagesConfig().giveCard()
-                    .replace("%player%", target.getName())
-                    .replace("%card%", rarity + " " + cardName));
-
-            target.getInventory().addItem(card.build(shiny));
+            ChatUtil.sendPrefixedMessage(sender, plugin.getMessagesConfig().givePack().replace("%player%", player.getName()).replace("%pack%", pack));
+            ChatUtil.sendPrefixedMessage(player, plugin.getMessagesConfig().boosterPackMsg());
         }
 
-        @Subcommand("shiny")
-        @CommandPermission(Permissions.GIVE_CARD_SHINY)
-        @CommandCompletion("@rarities @cards")
-        @Description("Gives a shiny card.")
-        public void onShiny(final Player player, @Single final String rarity, @Single final String cardName) {
-            onPlayer(player, player.getName(), rarity, cardName, true);
-        }
-    }
-
-    @Subcommand("pack")
-    @Description("Gives a pack to a player.")
-    @CommandCompletion("@players @packs")
-    @CommandPermission(Permissions.GIVE_PACK)
-    public void onGiveBoosterPack(final CommandSender sender, @Single final String playerName, @Single final String pack) {
-        Player player = Bukkit.getPlayerExact(playerName);
-        if (isOnline(player)) {
-            ChatUtil.sendPrefixedMessage(sender, CardsCommand.PLAYER_NOT_ONLINE);
-            return;
+        private boolean isOnline(final Player player) {
+            return player == null;
         }
 
-        CardUtil.dropItem(player, plugin.getPackManager().getPackItem(pack));
+        @Subcommand("random entity")
+        @Description("Gives a random card to a player.")
+        @CommandPermission(Permissions.GIVE_RANDOM_ENTITY)
+        public void onGiveRandomCard(final CommandSender sender, @Single final String playerName, final EntityType entityType) {
+            Player player = Bukkit.getPlayerExact(playerName);
+            if (isOnline(player)) {
+                ChatUtil.sendPrefixedMessage(sender, CardsCommand.PLAYER_NOT_ONLINE);
+                return;
+            }
 
-        ChatUtil.sendPrefixedMessage(sender, plugin.getMessagesConfig().givePack().replace("%player%", player.getName()).replace("%pack%", pack));
-        ChatUtil.sendPrefixedMessage(player, plugin.getMessagesConfig().boosterPackMsg());
-    }
-
-    private boolean isOnline(final Player player) {
-        return player == null;
-    }
-
-    @Subcommand("random entity")
-    @Description("Gives a random card to a player.")
-    @CommandPermission(Permissions.GIVE_RANDOM_ENTITY)
-    public void onGiveRandomCard(final CommandSender sender, @Single final String playerName, final EntityType entityType) {
-        Player player = Bukkit.getPlayerExact(playerName);
-        if (isOnline(player)) {
-            ChatUtil.sendPrefixedMessage(sender, CardsCommand.PLAYER_NOT_ONLINE);
-            return;
+            try {
+                String rare = plugin.getCardManager().getRandomRarity(CardUtil.getMobType(entityType), true);
+                plugin.debug(getClass(), "Rarity: " + rare);
+                ChatUtil.sendPrefixedMessage(sender, plugin.getMessagesConfig().giveRandomCardMsg().replace("%player%", player.getName()));
+                CardUtil.dropItem(player, plugin.getCardManager().getRandomCard(rare).build(false));
+            } catch (IllegalArgumentException exception) {
+                ChatUtil.sendPrefixedMessage(player, plugin.getMessagesConfig().noEntity());
+            }
         }
 
-        try {
-            String rare = plugin.getCardManager().getRandomRarity(CardUtil.getMobType(entityType), true);
-            plugin.debug(getClass(), "Rarity: " + rare);
-            ChatUtil.sendPrefixedMessage(sender, plugin.getMessagesConfig().giveRandomCardMsg().replace("%player%", player.getName()));
-            CardUtil.dropItem(player, plugin.getCardManager().getRandomCard(rare).build(false));
-        } catch (IllegalArgumentException exception) {
-            ChatUtil.sendPrefixedMessage(player, plugin.getMessagesConfig().noEntity());
-        }
-    }
+        @Subcommand("random rarity")
+        @Description("Gives a random card to a player. Specify rarity.")
+        @CommandCompletion("@players @rarities")
+        @CommandPermission(Permissions.GIVE_RANDOM_RARITY)
+        public void onGiveRandomCard(final CommandSender sender, @Single final String playerName, @Single final String rarity) {
+            Player player = Bukkit.getPlayerExact(playerName);
+            if (isOnline(player)) {
+                ChatUtil.sendPrefixedMessage(sender, CardsCommand.PLAYER_NOT_ONLINE);
+                return;
+            }
 
-    @Subcommand("random rarity")
-    @Description("Gives a random card to a player. Specify rarity.")
-    @CommandCompletion("@players @rarities")
-    @CommandPermission(Permissions.GIVE_RANDOM_RARITY)
-    public void onGiveRandomCard(final CommandSender sender, @Single final String playerName, @Single final String rarity) {
-        Player player = Bukkit.getPlayerExact(playerName);
-        if (isOnline(player)) {
-            ChatUtil.sendPrefixedMessage(sender, CardsCommand.PLAYER_NOT_ONLINE);
-            return;
-        }
-
-        try {
-            plugin.debug(GiveCommands.class,"Rarity: " + rarity);
-            ChatUtil.sendPrefixedMessage(sender, plugin.getMessagesConfig().giveRandomCardMsg().replace("%player%", player.getName()));
-            CardUtil.dropItem(player, plugin.getCardManager().getRandomCard(rarity).build(false));
-        } catch (IllegalArgumentException exception) {
-            ChatUtil.sendPrefixedMessage(player, plugin.getMessagesConfig().noEntity());
+            try {
+                plugin.debug(GiveCommands.class, "Rarity: " + rarity);
+                ChatUtil.sendPrefixedMessage(sender, plugin.getMessagesConfig().giveRandomCardMsg().replace("%player%", player.getName()));
+                CardUtil.dropItem(player, plugin.getCardManager().getRandomCard(rarity).build(false));
+            } catch (IllegalArgumentException exception) {
+                ChatUtil.sendPrefixedMessage(player, plugin.getMessagesConfig().noEntity());
+            }
         }
     }
 }
